# api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: Go Tasks API
  description: |
    Task and habit management API. Users register and manage their own tasks,
    categories, daily logs, and answers. All data is scoped to the authenticated user.

    ## Authentication

    Protected endpoints require a valid JWT access token in the `Authorization` header
    using the Bearer scheme. Access tokens are RS256-signed and valid for 15 minutes.

    **Token Claims:**
    - `sub` — user ID (UUID)
    - `iss` — issuer (configured via JWT_ISSUER)
    - `aud` — audience (configured via JWT_AUDIENCE)
    - `exp` — expiration time
    - `nbf` — not before time
    - `iat` — issued at time
    - `jti` — unique token ID (UUID)

    ## Refresh Tokens

    Refresh tokens are valid for 1 hour and support rotation. On use, the old token
    is deleted and a new one is issued. Logout revokes the refresh token and adds
    the access token's jti to a blocklist.

  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      description: |
        Rotates refresh token and issues new access token.
        Send the refresh token in the `X-Refresh-Token` header.
      operationId: refreshToken
      security: []
      parameters:
        - name: X-Refresh-Token
          in: header
          required: true
          description: Current refresh token
          schema:
            type: string
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponse"
        "401":
          description: Invalid or expired refresh token
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256-signed JWT access token
...
  schemas:
...
    RefreshResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: string
          format: date-time
          description: When the access token expires
        token_type:
          type: string
          example: Bearer
...